+++ src/DonorDashboards/Tabs/EditProfileTab/Tab.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Give\DonorDashboards\Tabs\EditProfileTab;
+
+use Give\DonorDashboards\Tabs\Contracts\Tab as TabAbstract;
+use Give\DonorDashboards\Tabs\EditProfileTab\ProfileRoute;
+use Give\DonorDashboards\Tabs\EditProfileTab\LocationRoute;
+
+/**
+ * @since 2.10.0
+ */
+class Tab extends TabAbstract {
+
+	public static function id() {
+		return 'edit-profile';
+	}
+
+	public function routes() {
+		return [
+			ProfileRoute::class,
+			LocationRoute::class,
+		];
+	}
+}